Another big update.  This should be ready for demoing.. ish.
I refactored a lot of things.  All page xml files now have the prefix "page_", to differentiate them from the dialog xml files (which begin with "dialog_").  There are 7 dialog xml files, but they are all used in one Dialog (which, in this case, is simply an Activity shown with a certain style).  

The PFD page will now show this dialog instead of what we had before with it giving a Toast message to say which food group it is in. These keep track of the food groups while the app is running.  If the app shuts off, a new Day object is created and all shares are forgotten.  This is because I haven't implemented the database yet.  This is the next step, but I don't think that we need it for the demo.  I don't know what Marcia wants specifically for this Dialog, so I didn't want to go too in-depth with database stuff if she was going to make us change and redo everything.

On that note, the History page is still completely blank.. since we'll find out tomorrow I guess what she wants for that.

I removed the Book java and xml files and replaced it with a simple link to the Amazon page for Life Preserver Diet.

The Skype link on the Contact page works correctly also.

Oh, I made a new graphic for the home screen.. I hope you don't mind, Devin.  It looks like a crappy imitation of what you'd made.  Perhaps you could edit it and make it look better at some point? If at all possible.

So, things left to do are:
-History page.. all of it
-Data storage and implementation with the PFD 
-Gender details for PFD (which may require database storage for user information?)
-Clean-up.. I think we need to make everything look a lot better than it does now. Perhaps new backgrounds for each page (PFD should probably have a background that looks like its floating in water or something).
-Splash screen - I still think it should have an animated splash screen with the logo on it (which needs to be redesigned).
-Adjustments for other screen sizes.. once we finish everything else.. We can adjust layout files and make it fit for everyone nicely.  Maybe fix up some graphics so that they look better than simple resizes.