Big update.
Got a halfway decent weekly graph going. A few variable changes are all that's necessary to extend it to a month.

The chosen date again is still that of the Day at the top of the database. The week containing that Day is found and all existing Days of that week are retrieved from the database. The share values are then plotted as a line of value vs day of the week (no markers yet so it's just numbers, 0 for Monday, 1 for Tuesday, etc..). I've also got a legend on the side and some random colors for each plot line to let the user know what line corresponds to dairy, fruit, etc...