Added good and bad images for all the food groups.
Also, added implementation for it on the PFD. If you're over, it'll show a red X over the icon, if you're right on your requirements, it's got a green check mark.

I can't remember what else I added, if anything.  I think just this.  It took way too long haha.