Graphs are bigger.
David : See how the second graph's title is "Exercise Minutes"? I think we should set the first one's title to "Share Values" or something similar. That way we don't need a y-axis label.

The date at the top, which is the first graph's current title, can be taken and written above the top graph in a TextView or something. The String var title starts being created at line 200 of GraphView.java